+++ Block/Adminhtml/Rate/Edit.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace ETechFlow\ShippingTableRates\Block\Adminhtml\Rate;
 
+use Magento\Backend\Block\Widget\Context;
 use Magento\Backend\Block\Widget\Form\Container;
+use Magento\Framework\Registry;
 
 /**
  * Container for the rate-rule edit form. Provides Save / Save and
@@ -12,6 +14,20 @@ use Magento\Backend\Block\Widget\Form\Container;
  */
 class Edit extends Container
 {
+    /**
+     * @var Registry
+     */
+    protected $_coreRegistry = null;
+
+    public function __construct(
+        Context $context,
+        Registry $registry,
+        array $data = []
+    ) {
+        $this->_coreRegistry = $registry;
+        parent::__construct($context, $data);
+    }
+
     /**
      * @return void
      */
@@ -43,7 +59,6 @@ class Edit extends Container
             $this->buttonList->remove('delete');
         }
 
-        // Override Back button to return to the parent method's edit page
         $method = $this->_coreRegistry->registry('etechflow_str_method');
         if ($method && $method->getMethodId()) {
             $backUrl = $this->getUrl('etechflow_str/method/edit', ['method_id' => $method->getMethodId()]);